refactor persistency core component
